hot fix step0_processor_s2_sr.py
Sometimes a tile is available twice with different processing times. This has led to an error when not all tiles were available, but another tile was available twice for a certain date. This resulted in "enough" tiles solely by number although not all of the area was covered. The "number" of tiles is now determined by disting tile id.