Update satromo_publish.py
Code adjustments to deal with product names with and without the prefix 'ch.swisstopo.' Including a new function get_collection_name. Thus had to differentiate between collection and product name.
Also added the scaling factor as it was called upon in the get_product_info function for the export of the warnregion files.